requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
numpy==2.1.3
